+++ src/providers/twitch/TwitchEmotes.cpp
@@ -468,8 +468,7 @@ EmotePtr TwitchEmotes::getOrCreateEmote(const EmoteId &id,
                                baseSize * (1.0 / emote3xScaleFactor)),
             },
             Tooltip{name.toHtmlEscaped() + "<br>Twitch Emote"},
-            Url{QString("https://emotes.awoo.nl/twitch/emote/%1")
-                    .arg(id.string)},
+            Url{QString("https://chatvau.lt/emote/twitch/%1").arg(id.string)},
         });
     }