+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Dankerino
+
+- Minor: Replace `emotes.awoo.nl` with `chatvau.lt`. (#147)
+
 ## Unversioned
 
 - Bugfix: Fixed crashes that could occur when Lua functions errored with values other than strings. (#6441)